ContentProvider: Group member count is wrong
When try to get the member count of a group, provider will
return the count of raw contacts in this group. If contact
is deleted, its raw contacts' deleted column will be marked
as 1. However, provider counts the 'deleted' raw contacts as
well.

We should return the count of contacts in this group in
stead of the count of raw contacts in this group.

CRs-Fixed: 752418

Change-Id: Ib65b66b44ad942a52ae1d3c24ff5c036462b0135