ContactsProvider: Label for Chinese polyphone is not right
When try to get bucket index, ICU will not consider the
Chinese polyphone. As a result, if the contact's first
name is Chinese polyphone, ICU can not return the right
index for the Chinese polyphone.

If current locale is Chinese, we should transliterate it
before we try to get the bucket index of the name.

CRs-Fixed: 620431

Change-Id: I138bf4065b6b033f00e98e9dd66cb28542ebab8e